% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf_appraise.R
\name{hf_appraise}
\alias{hf_appraise}
\title{Travel cost between points}
\usage{
hf_appraise(x, from, to)
}
\arguments{
\item{x}{a cost \code{terrain} generated by \code{hf_terrain()}.}

\item{from}{an \code{sf} specifying start locations, must be POINT or MULTIPOINT.}

\item{to}{an \code{sf} specifying end locations, must be POINT or MULTIPOINT.}
}
\value{
a \code{data.frame} with three columns: from, to, and cost. Values
in the from- and to-columns are \code{1:nrow(from)} and \code{1:nrow(to)}, respectively.
}
\description{
Estimate the cost of traveling the "shortest" or least-cost path between
multiple start and end points. For each from-point, the cost of traveling
the shortest path to every to-point is calculated.
}
\examples{
library(sf)
library(terra)

fn <- system.file("extdata/red_butte_dem.tif", package = "hiker")
red_butte_dem <- rast(fn)

from <- st_sf(geometry = st_sfc(st_point(c(432000, 4514000)),
                                crs = 26912))

to <- st_sf(geometry = st_sfc(st_point(c(431000, 4515000)),
                              st_point(c(436500, 4518500)),
                              crs = 26912))

terrain <- hf_terrain(red_butte_dem)

hf_appraise(terrain, from, to)

}
