% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf_barrier.R
\name{hf_barrier}
\alias{hf_barrier}
\title{Obstacle to travel}
\usage{
hf_barrier(x, barrier)
}
\arguments{
\item{x}{a cost \code{terrain} generated by \code{hf_terrain()}.}

\item{barrier}{an \code{sf} object.}
}
\value{
a \code{terrain} representing cost of travel.
}
\description{
Prevents travel along paths that cross through barriers or other obstacles that
cannot be traversed, like a river, a Greek phalanx, or
"\href{https://en.wikipedia.org/wiki/World_of_A_Song_of_Ice_and_Fire#The_Wall}{the Wall}".
}
\examples{
library(sf)
library(terra)

fn <- system.file("extdata/red_butte_dem.tif", package = "hiker")
red_butte_dem <- rast(fn)

fn <- system.file("extdata/red_butte_reservoir.geojson", package = "hiker")
red_butte_reservoir <- read_sf(fn)

terrain <- hf_terrain(red_butte_dem)

terrain <- hf_barrier(terrain, red_butte_reservoir)

plot(terrain, main = "Travel Cost")
plot(st_geometry(red_butte_reservoir),
     col = NA,
     border = "red2",
     add = TRUE)

}
