% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf_channel.R
\name{hf_channel}
\alias{hf_channel}
\title{Channel of reduced cost}
\usage{
hf_channel(x, channel, .m = 1)
}
\arguments{
\item{x}{a cost \code{terrain} generated by \code{hf_terrain()}.}

\item{channel}{an \code{sf} object.}

\item{.m}{a numeric between 0 and 1 applied as a weight to the cost of travel.}
}
\value{
a \code{terrain} representing cost of travel.
}
\description{
Reduces cost of travel along paths that cross through channels or
other aspects of a terrain that expedite travel. Examples include
hiking trails, sidewalks, and Einstein-Rosen bridges (otherwise
known as \href{https://en.wikipedia.org/wiki/Stargate_(device)}{stargates}).
}
\examples{
library(sf)
library(terra)

fn <- system.file("extdata/red_butte_dem.tif", package = "hiker")
red_butte_dem <- rast(fn)

fn <- system.file("extdata/red_butte_road.geojson", package = "hiker")
red_butte_road <- read_sf(fn)

terrain <- hf_terrain(red_butte_dem, hf = "tobler")

terrain <- hf_channel(terrain,
                      channel = red_butte_road,
                      .m = 0.6)

plot(terrain, main = "Travel Cost")
plot(st_geometry(red_butte_road),
     col = "red2",
     add = TRUE)

}
