% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf_hike.R
\name{hf_hike}
\alias{hf_hike}
\title{Shortest path between points}
\usage{
hf_hike(x, from, to, add_cost = FALSE)
}
\arguments{
\item{x}{a cost \code{terrain} generated by \code{hf_terrain()}.}

\item{from}{an \code{sf} specifying start locations, must be POINT or MULTIPOINT.}

\item{to}{an \code{sf} specifying end locations, must be POINT or MULTIPOINT.}

\item{add_cost}{logical, whether to include a column of travel costs.}
}
\value{
An \code{sf} object with least-cost \code{LINE} features and two columns: from and to.
Values in the from- and to-columns are \code{1:nrow(from)} and \code{1:nrow(to)}, respectively.
If \code{add_cost = TRUE}, the \code{sf} object also includes a travel cost column.
}
\description{
Calculate the "shortest" or least-cost paths between multiple start and end points.
For each from-point, the shortest path to every to-point is calculated.
}
\examples{
library(sf)
library(terra)

fn <- system.file("extdata/red_butte_dem.tif", package = "hiker")
red_butte_dem <- rast(fn)

from <- st_sf(geometry = st_sfc(st_point(c(432000, 4514000)),
                                st_point(c(434000, 4518000)),
                                crs = 26912))

to <- st_sf(geometry = st_sfc(st_point(c(431000, 4515000)),
                              st_point(c(436500, 4518500)),
                              crs = 26912))

terrain <- hf_terrain(red_butte_dem)

short_paths <- hf_hike(terrain, from, to)

plot(red_butte_dem)
plot(st_geometry(short_paths), lty = 2, add = TRUE)
plot(st_geometry(from), pch = 19, col = "red2", add = TRUE)
plot(st_geometry(to), pch = 19, col = "dodgerblue4", add = TRUE)

}
