% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf_rasterize.R
\name{hf_rasterize}
\alias{hf_rasterize}
\title{Terrain to raster}
\usage{
hf_rasterize(x)
}
\arguments{
\item{x}{a cost \code{terrain} generated by \code{hf_terrain()}.}
}
\value{
a \code{SpatRaster} with travel cost values.
}
\description{
Converts a cost \code{terrain} into a raster representing
travel cost, specifically the average cost of moving from/to each
cell. Mostly useful for visualizing with \code{terra::plot()}
as a sanity check.
}
\examples{
library(terra)

fn <- system.file("extdata/red_butte_dem.tif", package = "hiker")
red_butte_dem <- rast(fn)

terrain <- hf_terrain(red_butte_dem)

r_terrain <- hf_rasterize(terrain)

}
