% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf_survey.R
\name{hf_survey}
\alias{hf_survey}
\title{Accumulated travel time}
\usage{
hf_survey(x, from)
}
\arguments{
\item{x}{a cost \code{terrain} generated by \code{hf_terrain()}.}

\item{from}{an \code{sf} specifying start locations, must be POINT or MULTIPOINT.}
}
\value{
a \code{SpatRaster} with accumulated travel costs from each start point.
}
\description{
Survey entire landscape to calculate accumulated cost of travel from one or
more start points.
}
\examples{
library(sf)
library(terra)

fn <- system.file("extdata/red_butte_dem.tif", package = "hiker")
red_butte_dem <- rast(fn)

from <- st_sf(geometry = st_sfc(st_point(c(432000, 4514000)),
                                st_point(c(434000, 4518000)),
                                crs = 26912))

terrain <- hf_terrain(red_butte_dem)

cost_surface <- hf_survey(terrain, from)

plot(cost_surface)
plot(st_geometry(from), pch = 19, col = "red2", add = TRUE)

}
