% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf_terrain.R
\name{hf_terrain}
\alias{hf_terrain}
\alias{print.terrain}
\title{Cost surface}
\usage{
hf_terrain(x, hf = "campbell", max_slope = 45, neighbors = 8, ...)

\method{print}{terrain}(x, ...)
}
\arguments{
\item{x}{a \code{SpatRaster} with elevation values.}

\item{hf}{a character string specifying the preferred hiking function. Options
include \code{tobler} and \code{campbell} (default).}

\item{max_slope}{a numeric value specifying the maximum allowed hiking slope
\emph{in degrees} (default is 45). Suppose you ask: Would a person really hike
\emph{that}? This is the shallowest slope where the answer to that question is No.}

\item{neighbors}{numeric, one of 4, 8, or 16 (default) specifying the number of
adjacent cells between which to calculate movement cost. These are passed on
to \code{terra::adjacent()} and represent standard neighborhoods (rook, queen, and
knight + queen).}

\item{...}{Arguments passed on to the hiking function specified by \code{hf}.}
}
\value{
a \code{terrain} representing cost of travel.
}
\description{
Create a \code{terrain} representing the cost of travel. No assumption is
made about existing trails or paths.
}
\details{
The \code{SpatRaster} must have a projected CRS, with both distance units and elevation
values in meters.

For a list of available hiking functions and their parameters, see
\code{vignette("hiking functions", package = "hiker")}.
}
\examples{
library(terra)

fn <- system.file("extdata/red_butte_dem.tif", package = "hiker")
red_butte_dem <- rast(fn)

terrain <- hf_terrain(red_butte_dem)

# convert to raster to visualize
# plot(hf_rasterize(terrain), main = "Travel Cost")

# tobler on path
# terrain <- hf_terrain(red_butte_dem, hf = "tobler", off.path = FALSE)

# max slope
# terrain <- hf_terrain(red_butte_dem, max_slope = 25)

# neighbors
# terrain <- hf_terrain(red_butte_dem, neighbors = 16)

}
\references{
Campbell, Michael J., Philip E. Dennison, Bret W. Butler, and Wesley G. Page
(2019). "Using crowdsourced fitness tracker data to model the relationship
between slope and travel rates. \emph{Applied Geography} 106, 93-107.
\doi{https://doi.org/10.1016/j.apgeog.2019.03.008}

Tobler, Waldo R. (1993). "Three Presentations on Geographical Analysis
and Modeling: Non-Isotropic Geographic Modeling, Speculations
on the Geometry of Geography, and Global Spatial Analysis."
\emph{National Center for Geographic Information and Analysis, Technical Report} 93-1.
}
