% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike_create_lines.R
\name{hike_create_lines}
\alias{hike_create_lines}
\title{create_lines function}
\usage{
hike_create_lines(data)
}
\arguments{
\item{df}{with coords}
}
\value{
create a sf linestring from multiple input coords
}
\description{
create_lines function
}
\examples{
a <- stplanr::geo_code("Jena")
b <- stplanr::geo_code("Weimar")
a_to_b <- data.frame(rbind(a,b))
line <- hikeR::hike_create_lines(a_to_b)
class(line)
}
\author{
Eric Krueger
}
