% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike_create_lines.R
\name{hike_distance}
\alias{hike_distance}
\title{distance}
\usage{
hike_distance(point_a, point_b, unit)
}
\arguments{
\item{a}{to b, unit}
}
\value{
calc. the distance between to points
}
\description{
distance
}
