% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike_create_lines.R
\name{hike_height_diff}
\alias{hike_height_diff}
\title{height difference}
\usage{
hike_height_diff(elev_points, col)
}
\arguments{
\item{col}{- column with elevation}

\item{points}{- sf points}
}
\value{
calc. height difference in m
}
\description{
height difference
}
\examples{
a <- stplanr::geo_code("Jena")
b <- stplanr::geo_code("Weimar")
a_to_b <- data.frame(rbind(a,b))
elev_p <- hikeR::hike_spatial_elev(a_to_b,shiny_progress = F,apikey = Sys.getenv("apikey"))
hike_height_diff(elev_p,col = "elev")
}
