% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike_iso.R
\name{hike_iso2sf}
\alias{hike_iso2sf}
\title{trannsformation function --> geojson from ors to sf}
\usage{
hike_iso2sf(obj)
}
\arguments{
\item{geojson}{from ors pckg to sf (only isochrones)}
}
\value{
sf polygon
}
\description{
trannsformation function --> geojson from ors to sf
}
