% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike_iso.R
\name{hike_iso_create}
\alias{hike_iso_create}
\title{creates a iso chrone polygon for reachabilty display}
\usage{
hike_iso_create(x, y, range, profile)
}
\arguments{
\item{num}{coords, range, and profile(moving style)}
}
\value{
sf polygon
}
\description{
creates a iso chrone polygon for reachabilty display
}
