% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike_pKm.R
\name{hike_performance_km}
\alias{hike_performance_km}
\title{pKm - performance km}
\usage{
hike_performance_km(elev_points, col, tmp_route, routed)
}
\arguments{
\item{points}{with elev, col specify elev column}
}
\value{
performance km
}
\description{
pKm - performance km
}
