% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike_progressbar.R
\name{hike_progressBar}
\alias{hike_progressBar}
\title{renders a progressbar ui element}
\usage{
hike_progressBar(id, value, total = NULL, display_pct = FALSE,
  size = NULL, status = NULL, striped = FALSE, title = NULL)
}
\description{
renders a progressbar ui element
}
