% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike_routing.R
\name{hike_routing}
\alias{hike_routing}
\title{this function combines multiple routing providers and their different routing styles}
\usage{
hike_routing(data, shiny_progress, provider, profile, api)
}
\arguments{
\item{data}{frame of coords, shiny progress, profile and api}
}
\value{
return simple weather forecast by coords
}
\description{
routing
}
