% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike_scrape_pic.R
\name{hike_scrape_pic}
\alias{hike_scrape_pic}
\title{scrape wikipedia pictures by search term}
\usage{
hike_scrape_pic(searchTerm, maxItems)
}
\arguments{
\item{searchterm}{- string searchterm}

\item{maxItems-}{nr of max. pic. you want to look for, but it will stop if there is no more result}
}
\description{
scrape wikipedia pictures by search term
}
\examples{
hikeR::hike_scrape_pic("Grace Hopper")
}
