% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike_create_lines.R
\name{hike_search_plc}
\alias{hike_search_plc}
\title{search fun, which throws an error if typed in searhterm could not be found,
needed for shiny}
\usage{
hike_search_plc(instring)
}
\arguments{
\item{searchterm}{- string}
}
\value{
coords of searchterm
}
\description{
search fun, which throws an error if typed in searhterm could not be found,
needed for shiny
}
\examples{
 hike_search_plc("Jena")
}
