% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike_pKm.R
\name{hike_traveltime}
\alias{hike_traveltime}
\title{travel time in min and hours}
\usage{
hike_traveltime(pkm, speed)
}
\arguments{
\item{pkm}{data or nomral distance data, travelling speed}
}
\value{
travel time
}
\description{
travel time in min and hours
}
