% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike_warnings.R
\name{hike_warnungen}
\alias{hike_warnungen}
\title{this function retrives weather warnings from germany from the dwd}
\usage{
hike_warnungen()
}
\arguments{
\item{empty}{}
}
\value{
returns sf poly with weather warnings for germany
}
\description{
weather warnings
}
\author{
Patrick Schratz

Eric Krueger
}
