% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike_weather.R
\name{hike_weather}
\alias{hike_weather}
\title{weather}
\usage{
hike_weather(in_data)
}
\arguments{
\item{in_data}{- point coords to retrive weather info}
}
\value{
return simple weather forecast by coords
}
\description{
weather
}
\examples{
a <- stplanr::geo_code("Jena")
hikeR::hike_weather(a)
}
