% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\docType{class}
\name{HikeR-class}
\alias{HikeR-class}
\title{S4 class \code{HikeR}}
\description{
Formal class for classifying local minima and maxima
of a time series.
}
\section{Slots}{

\describe{
\item{\code{ys}}{\code{zoo}, time series with associated scores.}

\item{\code{k}}{\code{integer}, count of left/right neigbours around centre point.}

\item{\code{scoreby}}{\code{character}, scoring method.}

\item{\code{yname}}{\code{character}, name of the series.}
}}

