% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiker-package.R
\docType{package}
\name{hiker-package}
\alias{hiker}
\alias{hiker-package}
\title{hiker: Model Hiking Paths Over Costly Terrains}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The hiker package provides a suite of functions to model least-cost
    distances between points, least-cost paths between points, and accumulated 
    travel time from any point to any location across a landscape. To estimate 
    the cost of traversing a terrain, hiker implements a few well-known hiking 
    functions, including Tobler and Campbell.
}
\details{
See \code{vignette("hiker", package = "hiker")} for an overview of the package.
See \code{vignette("Hiking Functions", package = "hiker")} for the currently available
hiking functions and any additional arguments they may take.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kbvernon/hiker}
  \item Report bugs at \url{https://github.com/kbvernon/hiker/issues}
}

}
\author{
\strong{Maintainer}: Kenneth Blake Vernon \email{kenneth.b.vernon@gmail.com} (\href{https://orcid.org/0000-0003-0098-5092}{ORCID})

}
\keyword{internal}
