% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiker.R
\name{hiker}
\alias{hiker}
\title{Peak/trough scores of time series points}
\usage{
hiker(y, k, scoreby = c("vote", "avg", "diff", "max", "ent", "ttype",
  "hybrid"), tval = 1, confby = 3, ...)
}
\arguments{
\item{y}{\code{zoo}, univariate time series.}

\item{k}{\code{integer}, the count of left/right neighbours.}

\item{scoreby}{\code{character}, the scoring method to be used.}

\item{tval}{\code{numeric}, factor for standard deviation band
if \code{scoreby = 'ttype'}.}

\item{confby}{\code{integer}, count of minimum vote,
values in the set \code{3:5}.}

\item{...}{ellipsis argument.}
}
\value{
An object of S4-class \code{HikeR}.
}
\description{
This function computes the score value for each
data point of a time series. The first and last
\code{k} observations are set to \code{NA}.
}
\examples{
TEX <- SP500[, "TEX"]
ans <- hiker(TEX, k = 8, scoreby = "hybrid", tval = 0.1)
ans
#plot(ans)

}
\references{
Girish K. Palshikar. Simple Algorithms for
Peak Detection in Time-Series. In \emph{Proc. 1st Int. Conf.
Advanced Data Analysis,
Business Analytics and Intelligence}, 2009.
}
\seealso{
Other scores: \code{\link{score}}
}

