% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpdoc.R
\name{lpdoc}
\alias{lpdoc}
\title{Displaying the literate programming document.}
\usage{
lpdoc()
}
\description{
This function can be used for opening the literate programming pdf-file with
the default pdf-viewer.
}

