% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.terrain}
\alias{plot.terrain}
\title{Plot terrain}
\usage{
\method{plot}{terrain}(x, ...)
}
\arguments{
\item{x}{a terrain object}

\item{...}{additional arguments passed on to plot method for spatRaster}
}
\description{
Plot terrain
}
\details{
See terra's plot function for additional arguments.
}
\examples{
library(terra)

fn <- system.file("extdata/red_butte_dem.tif", package = "hiker")
red_butte_dem <- rast(fn)

terrain <- hf_terrain(red_butte_dem)

# plot(terrain, main = "Travel Cost")
}
