% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\docType{class}
\name{PTBB-class}
\alias{PTBB-class}
\title{S4 class \code{PTBB}}
\description{
Formal class for peaks, troughs, burst, busts and
intermittent phase of a time series.
}
\section{Slots}{

\describe{
\item{\code{pt}}{\code{zoo}, logical: indicating peak/trough points.}

\item{\code{type}}{\code{character}, type of point/phase.}

\item{\code{h}}{\code{numeric}, the threshhold for score evaluation.}
}}

