% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_butte_canyon_rna.R
\docType{data}
\name{rbc}
\alias{rbc}
\alias{red}
\alias{butte}
\alias{canyon,}
\alias{butte,}
\alias{rna}
\title{Red Butte Canyon Research Natural Area}
\format{
The DEm raster is saved as a TIF file, with elevation in meters.
The three vectors are saved as GEOJSON. All three vectors have a name
attribute and geometry. All spatial data share the same CRS,
\emph{NAD83 / UTM zone 12N} (EPSG: 26912).
}
\source{
Data for the DEM was downloaded from the US National Elevation Dataset
using the \href{https://docs.ropensci.org/FedData/}{\code{FedData}} package. The RNA
boundary polygon was downloaded from the
\href{https://data.fs.usda.gov/geodata/edw/datasets.php?dsetCategory=boundaries}{US Forest Service Geodata Clearinhouse}.
The RBC Reservoir polygon was (sadly) digitized by hand using Google Earth. And
the road linestring was derived from the
\href{https://gis.utah.gov/data/transportation/roads-system/}{Utah AGRC Roads and Highways}
dataset.
}
\description{
A spatial dataset for the Red Butte Canyon (RBC) Research Natural Area (RNA)
including a Digital Elevation Model (DEM), a polygon of the RNA boundary, a
a polygon of the RBC Reservoir, and a linestring of the RBC Road.
}
\details{
The vectors are all saved as external data rather than being exported by the package.
This is to avoid CRAN protesting about non-ASCII characters in their CRS.
}
\keyword{datasets}
