% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{scavgdiff}
\alias{scdiffmean}
\alias{scentropy}
\alias{schybrid}
\alias{scmaxdiff}
\alias{score}
\alias{scttype}
\alias{scvote}
\title{Basic scoring methods for local minima and maxima}
\usage{
score(x, k, scoreby = c("vote", "avg", "diff", "max", "ent", "ttype",
  "hybrid"), tval = 1, confby = 3, ...)

scmaxdiff(x, k)

scdiffmean(x, k)

scavgdiff(x, k)

scentropy(x, k, ...)

scttype(x, k, tval)

schybrid(x, k, tval, ...)

scvote(x, k, tval, confby = 3, ...)
}
\arguments{
\item{x}{\code{numeric}, vector of length \code{2 * k + 1}.}

\item{k}{\code{integer}, the count of left/right neighbours.}

\item{scoreby}{\code{character}, the scoring method to be used.}

\item{tval}{\code{numeric}, factor for standard deviation band
if \code{scoreby = 'ttype'}.}

\item{confby}{\code{integer}, count of minimum vote,
values in the set \code{3:5}.}

\item{...}{ellipsis argument.}
}
\value{
\code{numeric}, the score value.
}
\description{
These are basic functions for evaluating the centre
point of a time series as local minimum or maximum.
Hereby, a score value is computed according to various methods.
If the score is positive, the centre point is tentatively
classified as a local peak.
Incidentally, negative scores indicate a local minima.
}

