% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SP500}
\alias{SP500}
\title{Weekly price data of 476 S&P 500 constituents.}
\format{A S3-class \code{zoo} object with 265 weekly observations
of 476 members of the S&P 500 index. The sample starts at 2003-03-03
and ends in 2008-03-24.}
\source{
\url{http://w3.uniroma1.it/Tardella/datasets.html},\cr
\url{ http://finance.yahoo.com/}
}
\usage{
SP500
}
\description{
The data set was used in the reference below. The authors adjusted
the price data for dividends and have removed stocks if two or
more consecutive missing values were found. In the remaining cases
the NA entries have been replaced by interpolated values.
}
\references{
Cesarone, F. and Scozzari, A. and Tardella, F.: Portfolio
    selection problems in practice: a comparison between linear and
    quadratic optimization models, Working Paper, Universita degli
    Studi Roma Tre, Universita Telematica delle Scienze Umane and
    Universita di Roma, July 2010.
    \url{http://arxiv.org/ftp/arxiv/papers/1105/1105.3594.pdf}
}
\keyword{datasets}

