% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hikeR_app.R
\name{hike_app}
\alias{hike_app}
\title{run app}
\usage{
hike_app(elevation_api, cycle_api, graph_hopper_api, ors_api)
}
\arguments{
\item{elevation_api}{- google elec. https://developers.google.com/maps/documentation/elevation/#api_key}

\item{cycle_api}{- cyclestreets.net api}

\item{graph_hopper_api}{- optinal api (cause you have to pay for it)}

\item{ors_api}{- open route service api}
}
\description{
run app
}
<<<<<<< HEAD
\examples{
hikeR::hike_app(elevation_api,cycle_api,ors_api)
}
=======
>>>>>>> 75da34c43f2171faf0ab93b0851cf503ec0fc4b2
