% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike_create_lines.R
\name{hike_create_lines}
\alias{hike_create_lines}
\title{create lines
 @param df with coords
 @return create a line from multiple input coords
 @export}
\usage{
hike_create_lines(data)
}
\description{
create lines
 @param df with coords
 @return create a line from multiple input coords
 @export
}
