% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike_create_lines.R
\name{hike_height_diff}
\alias{hike_height_diff}
\title{height difference}
\usage{
hike_height_diff(elev_points, col)
}
\arguments{
\item{points}{with elev data, col = specify elev. data column}
}
\value{
calc. height difference in m
}
\description{
height difference
}
