% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hike_create_lines.R
\name{hike_search_plc}
\alias{hike_search_plc}
\title{redundant fun}
\usage{
hike_search_plc(instring)
}
\arguments{
\item{searchterm}{- string}
}
\value{
coords of searchterm
}
\description{
redundant fun
}
