% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R, R/HikerMethods.R
\docType{class}
\name{HikeR-class}
\alias{HikeR-class}
\alias{bursts}
\alias{bursts,HikeR-method}
\alias{busts}
\alias{busts,HikeR-method}
\alias{peaks}
\alias{peaks,HikeR-method}
\alias{phases}
\alias{phases,HikeR-method}
\alias{plot}
\alias{plot,HikeR,missing-method}
\alias{ridges}
\alias{ridges,HikeR-method}
\alias{show,HikeR-method}
\alias{summary}
\alias{summary,HikeR-method}
\alias{topeaks}
\alias{topeaks,HikeR-method}
\alias{totroughs}
\alias{totroughs,HikeR-method}
\alias{troughs}
\alias{troughs,HikeR-method}
\title{S4 class \code{HikeR}}
\usage{
\S4method{show}{HikeR}(object)

\S4method{summary}{HikeR}(object, ...)

\S4method{peaks}{HikeR}(object, h = 0)

\S4method{troughs}{HikeR}(object, h = 0)

\S4method{bursts}{HikeR}(object, h = 0, b = object@k)

\S4method{busts}{HikeR}(object, h = 0, b = object@k)

\S4method{ridges}{HikeR}(object, h = 0, b = object@k)

\S4method{phases}{HikeR}(object, h = 0, b = object@k)

\S4method{topeaks}{HikeR}(object, h = 0)

\S4method{totroughs}{HikeR}(object, h = 0)

\S4method{plot}{HikeR,missing}(x, type = c("series", "score", "both", "zoo"),
  h = 0, b = x@k, phase = c("none", "pt", "bb"), main = NULL,
  sub = NULL, pt.peak = list(col = "darkgreen", pch = 19, cex = 0.8),
  pt.trough = list(col = "darkred", pch = 19, cex = 0.8), area.se = list(col
  = "lightgray"), area.pb = list(col = "seagreen", density = 20),
  area.tb = list(col = "red2", density = 20), ...)
}
\arguments{
\item{object}{An object of S4 class \code{HikeR}.}

\item{...}{Ellipsis argument.}

\item{h}{\code{numeric}, the threshold value for scores
to be considered as peaks/troughs.}

\item{b}{\code{integer}, intermittent count of points between peaks.}

\item{type}{\code{character}, whether series, scores or both should be plotted.}

\item{phase}{\code{character}, whether burst/bust or topeak/totrough phases
should be drawn in series plot.}

\item{main}{\code{character}, main title of plot.}

\item{sub}{\code{character}, sub title of plot}

\item{pt.peak}{\code{list}, named elements are passed to \code{graphics::points()}.}

\item{pt.trough}{\code{list}, named elements are passed to \code{graphics::points()}.}

\item{area.se}{\code{list}, named elements are passed to \code{graphics::rect()}
for areas of pre- and post sample points.}

\item{area.pb}{\code{list}, named elements are passed to \code{graphics::rect()}
for 'to-peak' or 'burst' phases.}

\item{area.tb}{\code{list}, named elements are passed to \code{graphics::rect()}
for 'to-trough' or 'bust' phases.}
}
\value{
Object of S4-class \code{PTBB}.
}
\description{
Formal class for classifying local minima and maxima
of a time series.
}
\section{Slots}{

\describe{
\item{\code{ys}}{\code{zoo}, time series with associated scores.}

\item{\code{k}}{\code{integer}, count of left/right neigbours around centre point.}

\item{\code{scoreby}}{\code{character}, scoring method.}

\item{\code{yname}}{\code{character}, name of the series.}
}}

