% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R, R/HikerMethods.R
\docType{class}
\name{HikeR-class}
\alias{HikeR-class}
\alias{bursts}
\alias{bursts,HikeR-method}
\alias{busts}
\alias{busts,HikeR-method}
\alias{peaks}
\alias{peaks,HikeR-method}
\alias{ridges}
\alias{ridges,HikeR-method}
\alias{show,HikeR-method}
\alias{summary}
\alias{summary,HikeR-method}
\alias{troughs}
\alias{troughs,HikeR-method}
\title{S4 class \code{HikeR}}
\usage{
\S4method{show}{HikeR}(object)

\S4method{summary}{HikeR}(object, ...)

\S4method{peaks}{HikeR}(object, h = 0)

\S4method{troughs}{HikeR}(object, h = 0)

\S4method{bursts}{HikeR}(object, h = 0, b = object@k)

\S4method{busts}{HikeR}(object, h = 0, b = object@k)

\S4method{ridges}{HikeR}(object, h = 0, b = object@k)
}
\arguments{
\item{object}{An object of S4 class \code{HikeR}.}

\item{...}{Ellipsis argument.}

\item{h}{\code{numeric}, the threshold value for scores
to be considered as peaks/troughs.}

\item{b}{\code{integer}, intermittent count of points between peaks.}
}
\value{
Object of S4-class \code{PTBB}.
}
\description{
Formal class for classifying local minima and maxima
of a time series.
}
\section{Slots}{

\describe{
\item{\code{ys}}{\code{zoo}, time series with associated scores.}

\item{\code{k}}{\code{integer}, count of left/right neigbours around centre point.}

\item{\code{scoreby}}{\code{character}, scoring method.}

\item{\code{yname}}{\code{character}, name of the series.}
}}

